# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Podcast Platform API",
        "description": "API documentation for the Podcast Platform",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/episodes": {
            "get": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Get all episodes",
                "operationId": "7ec4c2e40b12c6630cbd49f872f4e1b0",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Filter episodes by title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "podcast_id",
                        "in": "query",
                        "description": "Filter episodes by podcast ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "featured",
                        "in": "query",
                        "description": "Filter episodes by featured status",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "Filter episodes published after this date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "Filter episodes published before this date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "published_at",
                                "title",
                                "duration_in_seconds"
                            ]
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Episode"
                                                    }
                                                },
                                                "links": {
                                                    "type": "object"
                                                },
                                                "meta": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Create a new episode",
                "operationId": "7065efab022e9b074c3a726627738807",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EpisodeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Episode created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Episode created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Episode"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        },
                                        "errors": {
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/episodes/featured": {
            "get": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Get featured episodes",
                "operationId": "e48e5eff44ec6de5a46e885824a5f2f9",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of episodes to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Episode"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/episodes/recent": {
            "get": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Get recent episodes",
                "operationId": "f50ddc7f120b32445389d3d3dd618ce6",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of episodes to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Episode"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/episodes/{slug}": {
            "get": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Get episode by slug",
                "operationId": "0f7f848180a0145e14938804bed254a1",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Episode slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Episode"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Episode not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Episode not found"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Episode": {
                "required": [
                    "id",
                    "title",
                    "slug",
                    "audio_url",
                    "duration_in_seconds",
                    "published_at",
                    "podcast_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "The Good, the Bad, and the Ugly"
                    },
                    "slug": {
                        "type": "string",
                        "example": "the-good-the-bad-and-the-ugly"
                    },
                    "description": {
                        "type": "string",
                        "example": "In this episode, we discuss..."
                    },
                    "audio_url": {
                        "type": "string",
                        "example": "https://example.com/audio/episode1.mp3"
                    },
                    "duration_in_seconds": {
                        "type": "integer",
                        "example": 1800
                    },
                    "formatted_duration": {
                        "type": "string",
                        "example": "30:00"
                    },
                    "transcript": {
                        "type": "string",
                        "example": "Full transcript of the episode..."
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": true
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "podcast": {
                        "$ref": "#/components/schemas/Podcast"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "EpisodeRequest": {
                "required": [
                    "title",
                    "slug",
                    "audio_url",
                    "duration_in_seconds",
                    "published_at",
                    "podcast_id"
                ],
                "properties": {
                    "podcast_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "The Good, the Bad, and the Ugly"
                    },
                    "slug": {
                        "type": "string",
                        "example": "the-good-the-bad-and-the-ugly"
                    },
                    "description": {
                        "type": "string",
                        "example": "In this episode, we discuss..."
                    },
                    "audio_url": {
                        "type": "string",
                        "example": "https://example.com/audio/episode1.mp3"
                    },
                    "duration_in_seconds": {
                        "type": "integer",
                        "example": 1800
                    },
                    "transcript": {
                        "type": "string",
                        "example": "Full transcript of the episode..."
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": true
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "API Token",
                "scheme": "bearer"
            }
        }
    },
...
}